add my information
beautify some component
