+++ frontend/package.json
@@ -1,4 +1,3 @@
-
 {
   "name": "refine-react-framework",
   "version": "0.1.0",
@@ -6,9 +5,8 @@
   "dependencies": {
     "@pankod/refine": "^2.5.0",
     "@pankod/refine-react-router": "^2.5.0",
-    "@pankod/refine-simple-rest": "^2.5.0",
+    "@pankod/refine-simple-rest": "^3.27.0",
     "@reduxjs/toolkit": "^1.8.3",
-    "@tanstack/react-query": "^4.0.10",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^12.0.0",
     "@testing-library/user-event": "^13.2.1",
