--- frontend/src/components/posts/index.jsx
@@ -1,3 +1,0 @@
-export * from "./CreatePost"
-export * from "./PostList"
-export * from "./EditPost"
