--- frontend/src/components/users/index.jsx
@@ -1,3 +1,0 @@
-export * from "./CreateUser"
-export * from "./UserList"
-export * from "./EditUser"
