+++ frontend/src/index.css
@@ -96,7 +96,13 @@
   color: white !important;
   font-size: 18px;
 }
+.ant-layout-sider-has-trigger {
+  min-width: 220px !important;
+  max-width: 220px !important;
+  width: 220px !important;
+}
 .ant-layout-sider-trigger {
+  width: 220px !important;
   background-color: #8fdc25 !important;
 }
 .ant-layout-sider-trigger:hover {
@@ -123,6 +129,6 @@
   height: 40px;
 }
 
-.ReactQueryDevtools {
+/* .ReactQueryDevtools {
   display: none !important;
-}
+} */
