+++ frontend/src/pages/users/UserService.jsx
@@ -1,5 +1,6 @@
-import axiosInStance, * as request from "common/AxiosInstance";
-import { store } from "../../redux/store";
+
+import axiosInStance, * as request from 'common/AxiosInstance';
+import { store } from '../../redux/store';
 import {
   deleteUserFailed,
   deleteUserStart,
@@ -7,19 +8,20 @@ import {
   updateUserFailed,
   updateUserStart,
   updateUserSuccess,
-} from "redux/usersSlice";
+} from 'redux/usersSlice';
 
-export const updateUser = async (userId, value, token, dispatch) => {
+export const updateUser = async (userid, value, dispatch) => {
   try {
     dispatch(updateUserStart());
+    const token = store.getState().auth.login.currentUser.access;
     const config = {
       headers: {
-        "content-type": "application/json; charset=utf-8",
+        'content-type': 'application/json; charset=utf-8',
         Authorization: `Bearer ${token}`,
       },
     };
     const res = await request.put(
-      `/user/update/${userId}`,
+      `/user/update/${userid}`,
       {
         email: value.email,
         username: value.username,
@@ -36,15 +38,19 @@ export const updateUser = async (userId, value, token, dispatch) => {
   }
 };
 
-export const deleteUser = async (userId, token, dispatch) => {
+export const deleteUser = async (userId, dispatch) => {
   try {
     dispatch(deleteUserStart());
+    const token = store.getState().auth.login.currentUser.access;
     const config = {
       headers: {
         Authorization: `Bearer ${token}`,
       },
     };
-    const res = await axiosInStance.delete(`/user/delete/${userId}`, config);
+    const res = await axiosInStance.delete(
+      `/user/delete/${userId}`,
+      config
+    );
     dispatch(deleteUserSuccess());
     return res;
   } catch (error) {
