+++ frontend/src/pages/users/index.jsx
@@ -1,0 +1,2 @@
+export * from './UserList';
+export * from './UserInfo';
